Implement "lambda-based" pattern matcher for COMFY macros.
This pattern-matcher uses anonymous functions to provide compiled
pattern-matching and expansion code for COMFY macros. One behavioral
change is that the resulting bindings are returned in the reverse
order of that produced by the original pattern-matching code.
That results in minor changes to some tests. Added additional tests
covering pattern matching and expansion.